std::to_string fix

+++ src/common/common.cc
@@ -6,16 +6,23 @@ std::string Instruction::str() const {
   std::stringstream ss{};
 
   ss << std::setw(9) << opTypeToString.at(type) << " {";
-  ss << "rs1 = " << std::setw(2) << std::to_string(rs1) << ", ";
-  ss << "rs2 = " << std::setw(2) << std::to_string(rs2) << ", ";
-  ss << "rs3 = " << std::setw(2) << std::to_string(rs3) << ", ";
+  ss << "rs1 = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(rs1)) << ", ";
+  ss << "rs2 = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(rs2)) << ", ";
+  ss << "rs3 = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(rs3)) << ", ";
 
-  ss << "rd = " << std::setw(2) << std::to_string(rd) << ", ";
-  ss << "rm = " << std::setw(2) << std::to_string(rm) << ", ";
-  ss << "csr = " << std::setw(2) << std::to_string(csr) << ", ";
+  ss << "rd = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(rd)) << ", ";
+  ss << "rm = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(rm)) << ", ";
+  ss << "csr = " << std::setw(2)
+     << std::to_string(static_cast<std::uint64_t>(csr)) << ", ";
 
   ss << "imm = 0x" << std::hex << std::setfill('0')
-     << std::setw(sizeof(imm) * 2) << std::to_string(imm) << "}";
+     << std::setw(sizeof(imm) * 2)
+     << std::to_string(static_cast<std::uint64_t>(imm)) << "}";
 
   return ss.str();
 }
